Work done on camera
Work done on camera so that it focuses on the player as they move around a level, via SFML's View class. Zoom function implemented, but will require work

New functions added to System Renderer that aloow the camera to move with the player, or focus on a specific vector.

Debug scene also added for testing purposes.

Engine modified so that when escape is pressed, it returns back to the main menu.